broke everything again
broke everything again, but this time I actually know what I am doing. The new config module has a much more flexible and stable system. "get_or" is the only way to access a field in the config, and it FORCES you to have a default value if there is not already one, and this default value is then saved to the config